--- packages/commonwealth/client/scripts/views/components/component_kit/CWBannerText/CWBannerText.scss
@@ -1,17 +1,0 @@
-.linkTextRow {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-.link {
-  color: #3498db;
-  text-decoration: none;
-  margin-right: 8px;
-  &:hover {
-    text-decoration: underline;
-  }
-}
-.text {
-  margin-right: 4px;
-}
