--- packages/commonwealth/client/scripts/views/components/component_kit/CWBannerText/index.ts
@@ -1,1 +1,0 @@
-export { CWBannerText as default } from './CWBannerText';
