+++ packages/commonwealth/client/scripts/views/components/component_kit/RenderTextWithLink/index.ts
@@ -1,0 +1,1 @@
+export { RenderTextWithLink as default } from './RenderTextWithLink';
